Fixes enemies

+++ Assets/Code/Weapon.cs
@@ -9,7 +9,7 @@ public class Weapon : ScriptableObject
 
   public Sprite artwork;
 
-  public float _damage;
+  public int _damage;
   public float _cooldown;
   public float _speed;
   public float _range;
